delete api file in client and rename some file names to start with single noun
