--- client/src/services/interceptorsSetup.js
@@ -1,2 +1,0 @@
-import axiosInstance from "./axios";
-import TokenService from "@/services/TokenService";
