+++ client/src/services/registrationService.js
@@ -1,5 +1,5 @@
 import { ENDPOINTS } from "@/config/api-routes";
-import api from "./axios";
+import api from "../axios/axios";
 
 export function register(nickname, email, password) {
   let registrationInfo = {
