+++ client/src/views/FlashcardsSet.vue
@@ -4,7 +4,7 @@
   </div>
 </template>
 <script>
-import SetService from "@/services/SetService";
+import SetService from "@/services/setService";
 import { flashcardsSetStore } from "@/store/flashcardsSetStore";
 
 export default {
