+++ flashcards-api/src/main/java/com/flashcardsapi/controllers/TagController.java
@@ -17,7 +17,7 @@ import lombok.AllArgsConstructor;
 @RestController
 @RequestMapping("tags")
 @AllArgsConstructor
-public class TagsController {
+public class TagController {
 
     private TagsService tagsService;
